# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SkillForge AI - Learning Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [student, instructor, admin]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /users/{userId}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update user role (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Categories
...
  /courses/{courseId}/reviews:
    get:
      operationId: getCourseReviews
      tags: [reviews]
      summary: Get reviews for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{reviewId}:
    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Update review
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete review
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Certificates
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [student, instructor, admin]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
        password:
          type: string
          minLength: 8
    UserRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [student, instructor, admin]
    UserListResponse:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Categories
...
    Review:
      type: object
      required: [id, courseId, userId, rating, createdAt]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatar:
          type: ["string", "null"]
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        instructorReply:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
    ReviewUpdate:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        instructorReply:
          type: string

    # Certificates
...